+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Login, and register with GitHub and GitLab OAuth #398
 - Configuration to turn off internal DockCI users #398
 - Links to GitHub repos #399
+- Basic parsing and folding of Docker build stage #400
 
 ### v0.0.9
 - Exception reported to Rollbar on git push hook `ValidationError` #322